Faking av http-respons treng vi fleire stadar
