+++ skribenten-backend/src/test/kotlin/no/nav/pensjon/brev/skribenten/services/KrrServiceTest.kt
@@ -1,15 +1,6 @@
 package no.nav.pensjon.brev.skribenten.services
 
-import com.fasterxml.jackson.databind.SerializationFeature
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.typesafe.config.ConfigFactory
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.mock.MockEngine
-import io.ktor.client.engine.mock.respond
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.http.HttpStatusCode
-import io.ktor.http.headersOf
-import io.ktor.serialization.jackson.jackson
 import kotlinx.coroutines.runBlocking
 import no.nav.pensjon.brev.skribenten.auth.FakeAuthService
 import no.nav.pensjon.brev.skribenten.services.KrrService.KontaktinfoKRRResponse
@@ -21,24 +12,6 @@ import kotlin.test.assertEquals
 
 class KrrServiceTest {
 
-    private val objectMapper = jacksonObjectMapper()
-
-    private fun settOppClient(body: Any): HttpClient =
-        HttpClient(MockEngine {
-            respond(
-                content = objectMapper.writeValueAsString(body),
-                status = HttpStatusCode.OK,
-                headers = headersOf("Content-Type", "application/json")
-            )
-
-        }) {
-            install(ContentNegotiation) {
-                jackson {
-                    disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-                }
-            }
-        }
-
     @Test
     fun `kan hente foretrukket spraak fra KRR`() {
         val respons = KontaktinfoKRRResponse(
@@ -49,7 +22,7 @@ class KrrServiceTest {
         val service = KrrService(
             config = ConfigFactory.parseMap(mapOf()),
             authService = FakeAuthService,
-            client = settOppClient(respons)
+            client = settOppHttpClient(respons)
         )
 
         runBlocking {
@@ -69,7 +42,7 @@ class KrrServiceTest {
         val service = KrrService(
             config = ConfigFactory.parseMap(mapOf()),
             authService = FakeAuthService,
-            client = settOppClient(respons)
+            client = settOppHttpClient(respons)
         )
 
         runBlocking {
@@ -89,7 +62,7 @@ class KrrServiceTest {
         val service = KrrService(
             config = ConfigFactory.parseMap(mapOf()),
             authService = FakeAuthService,
-            client = settOppClient(respons)
+            client = settOppHttpClient(respons)
         )
 
         runBlocking {
